gui: move system language inside setting app.
